fix cyclic dependency 1
